backend/requirements.txt
fastapi
uvicorn
sqlmodel
sse-starlette
qrcode
pillow
requests
python-dotenv

